Replace too generic BillingAddressConfiguration with parametrized BillingAddressMode
