+++ card/src/main/java/com/adyen/checkout/card/internal/ui/model/CardComponentParamsMapper.kt
@@ -8,7 +8,7 @@
 
 package com.adyen.checkout.card.internal.ui.model
 
-import com.adyen.checkout.card.BillingAddressConfiguration
+import com.adyen.checkout.card.BillingAddressMode
 import com.adyen.checkout.card.CardConfiguration
 import com.adyen.checkout.card.FieldMode
 import com.adyen.checkout.core.common.AdyenLogLevel
@@ -29,7 +29,6 @@ internal class CardComponentParamsMapper {
         paymentMethod: CardPaymentMethod?,
     ): CardComponentParams {
         val (commonComponentParams, sessionParams) = componentParamsBundle
-        val billingAddressConfiguration = cardConfiguration?.billingAddressConfiguration
         return CardComponentParams(
             commonComponentParams = commonComponentParams,
             showHolderName = cardConfiguration?.showHolderName ?: false,
@@ -39,8 +38,7 @@ internal class CardComponentParamsMapper {
             socialSecurityNumberMode = cardConfiguration?.socialSecurityNumberMode
                 ?: FieldMode.HIDE,
             koreanAuthenticationMode = cardConfiguration?.koreanAuthenticationMode ?: FieldMode.HIDE,
-            showPostalCode =
-                billingAddressConfiguration?.billingAddressMode is BillingAddressConfiguration.BillingAddressMode.PostalCode,
+            showPostalCode = cardConfiguration?.billingAddressMode is BillingAddressMode.PostalCode,
             cvcVisibility = if (cardConfiguration?.hideSecurityCode == true) {
                 CVCVisibility.ALWAYS_HIDE
             } else {
